Changed image uploader to file input, adjusted controller and views

+++ controllers/admin.js
@@ -1,5 +1,7 @@
-const Reunion = require('../models/reunion');
 const { validationResult } = require('express-validator');
+
+const Reunion = require('../models/reunion');
+const fileHelper = require('../util/file');
 
 exports.getReunions = (req, res, next) => {
   Reunion.find()
@@ -25,7 +27,28 @@ exports.getAddReunion = (req, res, next) => {
 };
 
 exports.postAddReunion = (req, res, next) => {
-  const { title, year, imageUrls, description } = req.body;
+  // const { title, year, imageUrls, description } = req.body;
+  const title = req.body.title;
+  const year = req.body.year;
+  const image = req.file;
+  const description = req.body.description;
+
+  if (!image) {
+    return res.status(422).render('admin/edit-reunion', {
+      pageTitle: 'Add Reunion',
+      path: '/admin/add-reunion',
+      editing: false,
+      hasError: true,
+      reunion: {
+        title: title,
+        year: year,
+        description: description,
+      },
+      errorMessage: 'Attached file is not a valid image',
+      validationErrors: errors.array(),
+    });
+  }
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -37,13 +60,14 @@ exports.postAddReunion = (req, res, next) => {
       reunion: {
         title: title,
         year: year,
-        imageUrls: imageUrls,
         description: description,
       },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array(),
     });
   }
+
+  const imageUrls = image.path;
 
   const reunion = new Reunion({
     title: title,
@@ -52,6 +76,7 @@ exports.postAddReunion = (req, res, next) => {
     description: description,
     userId: req.user,
   });
+  console.log(reunion);
   reunion
     .save()
     .then((result) => {
@@ -86,14 +111,18 @@ exports.getEditReunion = (req, res, next) => {
         reunion: reunion,
       });
     })
-    .catch((error) => console.log(error));
+    .catch((error) => {
+      const newError = new Error(error);
+      newError.httpStatusCode = 500;
+      return next(newError);
+    });
 };
 
 exports.postEditReunion = (req, res, next) => {
   const reunionId = req.body.reunionId;
   const updatedTitle = req.body.title;
   const updatedYear = req.body.year;
-  const updatedImageUrls = req.body.imageUrls;
+  const image = req.file;
   const updatedDescription = req.body.description;
   const errors = validationResult(req);
 
@@ -106,7 +135,6 @@ exports.postEditReunion = (req, res, next) => {
       product: {
         title: updatedTitle,
         year: updatedYear,
-        imageUrls: updatedImageUrls,
         description: updatedDescription,
         _id: reunionId,
       },
@@ -122,22 +150,40 @@ exports.postEditReunion = (req, res, next) => {
       }
       reunion.title = updatedTitle;
       reunion.year = updatedYear;
-      reunion.imageUrls = updatedImageUrls;
       reunion.description = updatedDescription;
+      if (image) {
+        fileHelper.deleteFile(reunion.imageUrls);
+        reunion.imageUrls = image.path;
+      }
       return reunion.save().then((result) => {
         console.log('Updated Reunion');
         res.redirect('/admin/reunions');
       });
     })
-    .catch((error) => console.log(error));
+    .catch((error) => {
+      const newError = new Error(error);
+      newError.httpStatusCode = 500;
+      return next(newError);
+    });
 };
 
 exports.postDeleteReunion = (req, res, next) => {
   const reunionId = req.body.reunionId;
-  Reunion.deleteOne({ _id: reunionId, userId: req.user._id })
+  Reunion.findById({ _id: reunionId, userId: req.user._id })
+    .then((reunion) => {
+      if (!reunion) {
+        return next(new Error('Reunion not found'));
+      }
+      fileHelper.deleteFile(product.imageUrls);
+      return Reunion.deleteOne({ _id: reunionId, userId: req.user._id });
+    })
     .then(() => {
       console.log('Reunion Deleted');
       res.redirect('/admin/reunions');
     })
-    .catch((error) => console.log(error));
+    .catch((error) => {
+      const newError = new Error(error);
+      newError.httpStatusCode = 500;
+      return next(newError);
+    });
 };
